+++ src/wallet/address_validator.rs
@@ -52,7 +52,7 @@
 //!     fn validate(
 //!         &self,
 //!         keychain: KeychainKind,
-//!         hd_keypaths: &HDKeyPaths,
+//!         bip32_derivation: &BIP32Derivation,
 //!         script: &Script
 //!     ) -> Result<(), AddressValidatorError> {
 //!         let address = Address::from_script(script, Network::Testnet)
@@ -60,7 +60,7 @@
 //!             .map(Address::to_string)
 //!             .unwrap_or(script.to_string());
 //!         println!("New address of type {:?}: {}", keychain, address);
-//!         println!("HD keypaths: {:#?}", hd_keypaths);
+//!         println!("HD keypaths: {:#?}", bip32_derivation);
 //!
 //!         Ok(())
 //!     }
@@ -79,7 +79,7 @@ use std::fmt;
 
 use bitcoin::Script;
 
-use crate::descriptor::HDKeyPaths;
+use crate::descriptor::BIP32Derivation;
 use crate::types::KeychainKind;
 
 /// Errors that can be returned to fail the validation of an address
@@ -117,7 +117,7 @@ pub trait AddressValidator: Send + Sync + fmt::Debug {
     fn validate(
         &self,
         keychain: KeychainKind,
-        hd_keypaths: &HDKeyPaths,
+        bip32_derivation: &BIP32Derivation,
         script: &Script,
     ) -> Result<(), AddressValidatorError>;
 }
@@ -135,7 +135,7 @@ mod test {
         fn validate(
             &self,
             _keychain: KeychainKind,
-            _hd_keypaths: &HDKeyPaths,
+            _bip32_derivation: &BIP32Derivation,
             _script: &bitcoin::Script,
         ) -> Result<(), AddressValidatorError> {
             Err(AddressValidatorError::InvalidScript)